Remove extension of IntegrationTest and remove the field taxRate from the test. Close #1318
